move rate limit error constant to cloud constants
