+++ ui/src/cloud/constants/index.ts
@@ -1,0 +1,4 @@
+export const RATE_LIMIT_ERROR_STATUS = 429
+
+export const RATE_LIMIT_ERROR_TEXT =
+  'Oops. It looks like you have exceeded the query limits allowed as part of your plan. If you would like to increase your query limits, reach out to cloudbeta@influxdata.com.'
