--- ui/src/shared/constants/errors.ts
@@ -1,1 +1,0 @@
-export const RATE_LIMIT_ERROR_STATUS = 429
